+++ app/src/main/java/com/izzed/indianflavors/model/Cart.kt
@@ -4,7 +4,7 @@ data class Cart(
     var id: Int? = null,
     var productId: Int? = null,
     val productName: String,
-    val productPrice: Double,
+    val productPrice: Int,
     val productImgUrl: String,
     var itemQuantity: Int = 0,
     var itemNotes: String? = null
